chore: clean duplicate import
